hopefully support browsers

+++ src/gi/pipeline_assets.rs
@@ -1,15 +1,17 @@
 use bevy::prelude::*;
-use bevy::render::render_resource::{StorageBuffer, UniformBuffer};
+use bevy::render::render_resource::UniformBuffer;
 use bevy::render::renderer::{RenderDevice, RenderQueue};
 use bevy::render::Extract;
 use rand::{thread_rng, Rng};
 
+use super::types_gpu::{
+    GpuLightOccluder2DArray, GpuOmniLightSourceArray, GpuProbeArray, GpuSkylightMaskArray,
+};
 use crate::gi::constants::GI_SCREEN_PROBE_SIZE;
 use crate::gi::resource::ComputedTargetSizes;
 use crate::gi::types::{LightOccluder2D, OmniLightSource2D, SkylightLight2D, SkylightMask2D};
 use crate::gi::types_gpu::{
-    GpuCameraParams, GpuLightOccluder2D, GpuLightOccluderBuffer, GpuLightPassParams,
-    GpuLightSourceBuffer, GpuOmniLightSource, GpuProbeDataBuffer, GpuSkylightMaskBuffer,
+    GpuCameraParams, GpuLightOccluder2D, GpuLightPassParams, GpuOmniLightSource,
     GpuSkylightMaskData,
 };
 use crate::prelude::BevyMagicLight2DSettings;
@@ -20,10 +22,10 @@ use crate::FloorCamera;
 pub(crate) struct LightPassPipelineAssets {
     pub camera_params:     UniformBuffer<GpuCameraParams>,
     pub light_pass_params: UniformBuffer<GpuLightPassParams>,
-    pub light_sources:     StorageBuffer<GpuLightSourceBuffer>,
-    pub light_occluders:   StorageBuffer<GpuLightOccluderBuffer>,
-    pub probes:            StorageBuffer<GpuProbeDataBuffer>,
-    pub skylight_masks:    StorageBuffer<GpuSkylightMaskBuffer>,
+    pub probes:            UniformBuffer<GpuProbeArray>,
+    pub light_sources:     UniformBuffer<GpuOmniLightSourceArray>,
+    pub light_occluders:   UniformBuffer<GpuLightOccluder2DArray>,
+    pub skylight_masks:    UniformBuffer<GpuSkylightMaskArray>,
 }
 
 impl LightPassPipelineAssets {
@@ -66,14 +68,15 @@ pub(crate) fn system_extract_pipeline_assets(
     *gpu_target_sizes = **res_target_sizes;
 
     {
-        let mut light_sources = gpu_pipeline_assets.light_sources.get_mut();
-        let mut rng = thread_rng();
+        let light_sources = gpu_pipeline_assets.light_sources.get_mut();
         light_sources.count = 0;
-        light_sources.data.clear();
+        let mut rng = thread_rng();
         for (transform, light_source, visibility) in query_lights.iter() {
             if visibility.is_visible() {
-                light_sources.count += 1;
-                light_sources.data.push(GpuOmniLightSource::new(
+                if light_sources.count as usize >= light_sources.data.len() {
+                    break;
+                }
+                light_sources.data[light_sources.count as usize] = GpuOmniLightSource::new(
                     OmniLightSource2D {
                         intensity: light_source.intensity
                             + rng.gen_range(-1.0..1.0) * light_source.jitter_intensity,
@@ -85,36 +88,41 @@ pub(crate) fn system_extract_pipeline_assets(
                         transform.translation.y
                             + rng.gen_range(-1.0..1.0) * light_source.jitter_translation,
                     ),
-                ));
+                );
+                light_sources.count += 1;
             }
         }
     }
 
     {
-        let mut light_occluders = gpu_pipeline_assets.light_occluders.get_mut();
+        let light_occluders = gpu_pipeline_assets.light_occluders.get_mut();
         light_occluders.count = 0;
-        light_occluders.data.clear();
         for (occluder, transform, visibility) in query_occluders.iter() {
             if visibility.is_visible() {
-                light_occluders.count += 1;
-                light_occluders.data.push(GpuLightOccluder2D::new(
+                if light_occluders.count as usize >= light_occluders.data.len() {
+                    break;
+                }
+                light_occluders.data[light_occluders.count as usize] = GpuLightOccluder2D::new(
                     transform,
                     occluder.h_size,
-                ));
+                );
+                light_occluders.count += 1;
             }
         }
     }
 
-    {
-        let mut skylight_masks = gpu_pipeline_assets.skylight_masks.get_mut();
+    {   
+        let skylight_masks = gpu_pipeline_assets.skylight_masks.get_mut();
         skylight_masks.count = 0;
-        skylight_masks.data.clear();
         for (transform, mask) in query_masks.iter() {
-            skylight_masks.count += 1;
-            skylight_masks.data.push(GpuSkylightMaskData::new(
+            if skylight_masks.count as usize >= skylight_masks.data.len() {
+                break;
+            }
+            skylight_masks.data[skylight_masks.count as usize] = GpuSkylightMaskData::new(
                 transform.translation.truncate(),
                 mask.h_size,
-            ));
+            );
+            skylight_masks.count += 1;
         }
     }
